Refactor UI | Update TaskDao.kt

+++ app/src/main/java/dev/spikeysanju/einsen/data/datastore/db/TaskDao.kt
@@ -1,10 +1,6 @@
 package dev.spikeysanju.einsen.data.datastore.db
 
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import androidx.room.Update
+import androidx.room.*
 import dev.spikeysanju.einsen.model.Task
 import kotlinx.coroutines.flow.Flow
 
@@ -18,11 +14,11 @@ interface TaskDao {
     suspend fun insertSource(task: Task)
 
     @Query("DELETE FROM task where id =:id")
-    suspend fun deleteByID(id: Long)
+    suspend fun deleteByID(id: Int)
 
     @Update(onConflict = OnConflictStrategy.REPLACE)
     suspend fun updateTask(task: Task)
 
     @Query("SELECT * FROM task where id=:id")
-    fun findByID(id: Long): Flow<Task>
+    fun findByID(id: Int): Flow<Task>
 }
