+++ app/src/main/java/dev/spikeysanju/einsen/repository/MainRepository.kt
@@ -15,6 +15,6 @@ class MainRepository @Inject constructor(private val taskDao: TaskDao) {
 
     suspend fun insert(task: Task) = taskDao.insertSource(task)
     suspend fun update(task: Task) = taskDao.insertSource(task)
-    suspend fun delete(id: Long) = taskDao.deleteByID(id)
-    fun find(id: Long) = taskDao.findByID(id).flowOn(Dispatchers.IO).conflate()
+    suspend fun delete(id: Int) = taskDao.deleteByID(id)
+    fun find(id: Int) = taskDao.findByID(id).flowOn(Dispatchers.IO).conflate()
 }
