+++ app/src/main/java/dev/spikeysanju/einsen/view/viewmodel/MainViewModel.kt
@@ -79,7 +79,7 @@ class MainViewModel @Inject constructor(private val repo: MainRepository) : View
     }
 
     // delete source
-    fun deleteTaskByID(id: Long) = viewModelScope.launch {
+    fun deleteTaskByID(id: Int) = viewModelScope.launch {
         repo.delete(id)
     }
 
@@ -89,7 +89,7 @@ class MainViewModel @Inject constructor(private val repo: MainRepository) : View
     }
 
     // find task by id
-    fun findTaskByID(id: Long) = viewModelScope.launch(Dispatchers.IO) {
+    fun findTaskByID(id: Int) = viewModelScope.launch(Dispatchers.IO) {
         repo.find(id).distinctUntilChanged().collect { result ->
             try {
                 if (result.title.isEmpty()) {
